Drop OPENSSL_cleanse asm implementations
