+++ Configure
@@ -429,7 +429,7 @@ my %table=(
     },
     aarch64_asm => {
 	template	=> 1,
-	cpuid_obj       => "armcap.o arm64cpuid.o mem_clr.o",
+	cpuid_obj       => "armcap.o arm64cpuid.o",
 	ec_obj          => "ecp_nistz256.o ecp_nistz256-armv8.o",
 	bn_obj          => "bn_asm.o armv8-mont.o",
 	aes_obj         => "aes_core.o aes_cbc.o aesv8-armx.o vpaes-armv8.o",
@@ -1660,7 +1660,6 @@ if ($fips)
 	$openssl_other_defines.="#define OPENSSL_FIPS\n";
 	}
 
-$cpuid_obj="mem_clr.o"	unless ($cpuid_obj =~ /\.o$/);
 $des_obj=$des_enc	unless ($des_obj =~ /\.o$/);
 $bf_obj=$bf_enc		unless ($bf_obj =~ /\.o$/);
 $cast_obj=$cast_enc	unless ($cast_obj =~ /\.o$/);
@@ -2016,7 +2015,7 @@ print OUT "#ifdef OPENSSL_ALGORITHM_DEFINES\n";
 print OUT $openssl_algorithm_defines_trans;
 print OUT "#endif\n\n";
 
-print OUT "#define OPENSSL_CPUID_OBJ\n\n" if ($cpuid_obj ne "mem_clr.o");
+print OUT "#define OPENSSL_CPUID_OBJ\n\n" if ($cpuid_obj);
 
 while (<IN>)
 	{
